+++ app/middlewares/user.go
@@ -1,0 +1,97 @@
+package middlewares
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/getfider/fider/app/pkg/validate"
+
+	"github.com/getfider/fider/app/models"
+
+	"github.com/getfider/fider/app"
+	"github.com/getfider/fider/app/pkg/errors"
+	"github.com/getfider/fider/app/pkg/jwt"
+	"github.com/getfider/fider/app/pkg/web"
+	"github.com/getfider/fider/app/pkg/web/util"
+)
+
+// User gets JWT Auth token from cookie and insert into context
+func User() web.MiddlewareFunc {
+	return func(next web.HandlerFunc) web.HandlerFunc {
+		return func(c web.Context) error {
+			var (
+				token string
+				user  *models.User
+			)
+
+			cookie, err := c.Request.Cookie(web.CookieAuthName)
+			if err == nil {
+				token = cookie.Value
+			} else {
+				token = webutil.GetSignUpAuthCookie(c)
+				if token != "" {
+					webutil.AddAuthTokenCookie(c, token)
+				}
+			}
+
+			if token != "" {
+				claims, err := jwt.DecodeFiderClaims(token)
+				if err != nil {
+					c.RemoveCookie(web.CookieAuthName)
+					c.SetClaims(claims)
+					return next(c)
+				}
+
+				user, err = c.Services().Users.GetByID(claims.UserID)
+				if err != nil {
+					if errors.Cause(err) == app.ErrNotFound {
+						c.RemoveCookie(web.CookieAuthName)
+						return next(c)
+					}
+					return err
+				}
+			} else {
+				authHeader := c.Request.GetHeader("Authorization")
+				parts := strings.Split(authHeader, "Bearer")
+				if len(parts) == 2 {
+					apiKey := strings.TrimSpace(parts[1])
+					user, err = c.Services().Users.GetByAPIKey(apiKey)
+					if err != nil {
+						if errors.Cause(err) == app.ErrNotFound {
+							return c.HandleValidation(validate.Failed("API Key is invalid"))
+						}
+						return err
+					}
+
+					if !user.IsCollaborator() {
+						return c.HandleValidation(validate.Failed("API Key is invalid"))
+					}
+
+					if impersonateUserIDStr := c.Request.GetHeader("X-Fider-UserID"); impersonateUserIDStr != "" {
+						if !user.IsAdministrator() {
+							return c.HandleValidation(validate.Failed("Only Administrators are allowed to impersonate another user"))
+						}
+						impersonateUserID, err := strconv.Atoi(impersonateUserIDStr)
+						if err != nil {
+							return c.HandleValidation(validate.Failed(fmt.Sprintf("User not found for given impersonate UserID '%s'", impersonateUserIDStr)))
+						}
+						user, err = c.Services().Users.GetByID(impersonateUserID)
+						if err != nil {
+							if errors.Cause(err) == app.ErrNotFound {
+								return c.HandleValidation(validate.Failed(fmt.Sprintf("User not found for given impersonate UserID '%s'", impersonateUserIDStr)))
+							}
+							return err
+						}
+					}
+				}
+			}
+
+			if user != nil && c.Tenant() != nil && user.Tenant.ID == c.Tenant().ID {
+				c.SetUser(user)
+			}
+
+			return next(c)
+		}
+	}
+}
